pyproject.toml
[tool.poetry]
name = "bclproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2.16"
djangorestframework = "^3.14.0"
qrcode = "^7.4.2"
pillow = "^10.0.0"
secure-smtplib = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
